Manually build the CSS with sass
The main reason I'm doing this as it avoids the need for a css loader. It does means that library consumers will have to import the CSS somewhere as well as the JS.

If we wanted a better name for the import (e.g 'mycrossword/index.css' rather than including the dist folder) we could place the css in the project root.

As part of this I've added concurrently so you can build the JS and CSS at the same time