build package (#14)
* Set up a dev container to make it easier to run project within vscode

* Export component instead of rendering it

Export the props interface so that we can use it from the app.

* Remove react-scripts and setup webpack

This temporarily breaks linting and tests

* Add in example that uses create-react-app

For testing the built project, which is useful for making sure build has worked and types are exported correctly. This won't live-reload.

* added jest

* Add typing to App

* Switch webpack for plain tsc

I found building with webpack painful - at each step
there was a new problem to solve. I struggled to get the types
built in a way that matched the bundled file webpack producted.

With a couple of modifications to the project we can use
tsc and don't need webpack.

* Remove base url

Typescript support relative URL but doesn't transpile them.
If we wanted to use them we would need to introduce an additional tool and step in the build.

See https://github.com/microsoft/TypeScript/issues/10866 for more.

* Link example to use built mycrossword package

I've also linked react and react-dom - without this there's two versions of react and you see the 'Invalid hook' react from react - https://reactjs.org/warnings/invalid-hook-call-warning.html#duplicate-react

* Type example

We could import a typed example, but this is more like the usage we expect in the real world - the consumer of the library will have their own data, and thay may want to type it.

* Manually build the CSS with sass

The main reason I'm doing this as it avoids the need for a css loader. It does means that library consumers will have to import the CSS somewhere as well as the JS.

If we wanted a better name for the import (e.g 'mycrossword/index.css' rather than including the dist folder) we could place the css in the project root.

As part of this I've added concurrently so you can build the JS and CSS at the same time

* Create new tsconfig for build to stop building test files

* Clean build folder before creating build

In case the build no longer has files used in the previous build

* Remove ts-loader

* Stop jest running tests in dist folder

* lint files

* removed public folder

* update readme

* removed eslint

Co-authored-by: t-blackwell <tlb.blackwell@googlemail.com>

Closes #13