Behaviour improvements around grouped clues (#56)
* re-enable eslint by ignoring jest config files

* set prettier linefeed preference to auto

git should normalise the line endings to the most suitable to the
platform (lfs on linux/mac, crlfs on windows), which means that when a
contributor tries to run the format script on a linux box, each line in
the repo gets rewritten.
if we use auto, then (hopefully!) prettier should maintain whichever
ending was checked out by git, and the format script remains usable for
all, regardless of platform

* Highlight cells belonging to clues in the selected group

It's a bit of an edge case, but it's totally valid for a clue to belong
to many groups. An example is Guardian Prize 25220 - the cells for
clue 22a belong to both the 22a,14a group and the 23a,25a,... group. By
assocating the cell with exactly one group, the group encountered last
wins, so when selecting a cell in 22a, 23a is highlighted instead of 14a
which one would expect.

As a clue entry can only declare a single group, we should trust that it
describes the "most relevant" group to that clue - ie. if the clue
appears at the start of a group then we should expect that group to be in
the entry's `group` field.

As such, a better way to find the cells to be highlighted is to test if
the cell's `clueIds` overlaps with the `selectedClue`'s `group`. Once we
do that, we no longer need to store a group for a cell, simplifying the
cell data model and initialisation process.

* Highlight clues belonging to same group as lead

Same idea as previous commit - when highlighting grouped clues, prefer
the group declared in the entry for the selected clue, as it usually
makes the most sense

* include prize 25220 sample

* reset default example crossword

* apply suggested isHighlighted improvement